Add bootstrapper
bootstrap_decoded_binary() is the function that copies the decoded image
over a running one. In order not to break the running code, it is first
copied to higher memory (hardcoded to be 0x30000 ATM) and called from
there. After it's done copying, it jumps to 0x8000 to start executing
the new code.

The buff and decodebuff have been changed to point to fixed addresses in
memory in order to be big enough to fit meaningful data.