Fix boot code, install interrupt vectors properly
Ditch everything there was in boot.S and move Start symbol to point to
interrupt vector table. Now the interrupt table starts at the very first
bytes of the executable image and reset_handler can hardcode its
address.

Note: the code from boot.S is useful and will be restored at some point
somewhere, but for now it was an impediment.