Unhardcode irq vector table address again
This time made sure it really works, albeit I have to admit it's pretty
mysterious...

Here's the story. If I simply make this change:

    -    mov     r0, #0x8000
    +    ldr     r0, =irq_vectors

and leave the surrounding code as is, the compiler (or linker) would
place the address of irq_vectors after irq_handler and refer to it via
[pc + offset] and it doesn't work. But if I move enable_irq and
irq_handler away to another file, the address of irq_vectors is placed
right after reset_handler and is referred to via [pc + smaller offset]
and THAT works. I've no idea why, it makes no sense.