Put all stacks near the end of RAM
Pretend we have 128 megs or RAM. Plenty for now and can be detected
later. Designate a generous 1meg for each stack: IRQ, FIQ, SVR and SYS.
User mode stack is not set now, will be set for each userland process
separately.

Note that this commit changes the mode in which the core is left after
reset: it used to stay in Supervisor mode, now it's in System mode.